Create max_element_in_array_after_sum.py
Given the sample question:

a b k
1 5 3
4 8 7
6 9 1

The resulting transformations to the array are
```
 1  2  3  4  5   6  7  8   9  10
[0, 0, 0, 0, 0,  0, 0, 0,  0, 0]
[3, 0, 0, 0, 0, -3, 0, 0,  0, 0] // a: 1, b: 5, k: 3, so b+1 = 6
[3, 0, 0, 7, 0, -3, 0, 0, -7, 0]// a: 4: , b: 8, k: 7, so b+1 = 9
[3, 0, 0, 7, 0, -2, 0, 0, -7, -1] // a: 6, b: 9, k: 1, so b+1 = 10
```

you iterate through the final state of the array, noting when the slope increase and comparing that value to the max value so far until you reach the end. In your example above, it you start at 3, then you see a 7, which adds to 10. All subsequent slopes are negative and so the max is 10.